// PropertyRenting.Api/ClientApp/src/app/Models/contact-person.ts
export interface ContactPerson {
    tempId?: any;
    id?: any;
    status?: boolean;
    code?: number;
    nameAR?: string;
    nameEN?: string;
    phone1?: string;
    mobile1?: string;
    email?: string;
    identityTypeId?: number;
    identityNumber?: string;
    identityIssuePlace?: string;
    identityIssueDate?: Date;
    identityExpiryDate?: Date;
    notes?: string;
    renterId?: any;
}